+++ HISTORY.rst
@@ -2,11 +2,22 @@
 History
 =======
 
+0.1.3 (2023-08-16)
+
+* BREAKING CHANGE: `Network` no longer accept settings. Individual setting are now argument for Network.
+* Bugg fixes.
+
+
+0.1.2 (2023-06-14)
+
+* `tensor` method for NetworkObject
+
+
 0.1.1 (2023-05-26)
 ------------------
 
 * Every NetworkObject can have a recorder behavior.
-* Netowrk settings accept "index" entry.
+* Network settings accept "index" entry.
 * Bug fixes and general improvement.